refactor: refactor product creation module
