--- src/components/templates/MultiPageFormBox.tsx
@@ -1,13 +1,0 @@
-import MultiPageForm from "../organisms/MultiPageForm";
-
-const ProductCreation = () => {
-  return (
-    <div className="min-h-screen flex flex-col gap-8 items-center justify-center">
-      <div className="w-4/5 md:w-3/5">
-        <MultiPageForm />
-      </div>
-    </div>
-  );
-};
-
-export default ProductCreation;
