+++ src/modules/product-creation/components/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductCreationForm";
