+++ src/shared/constants/app.constants.ts
@@ -1,0 +1,5 @@
+export const rentOptions = [
+  { value: "", label: "Select an option" },
+  { value: "hr", label: "per hr" },
+  { value: "day", label: "per day" },
+];
